+++ tests/test_convster/test_parallel_convster.py
@@ -1,27 +1,25 @@
-import builtins
-from functools import partial
-
-from time import sleep
-
-import numpy as np
-import multiprocessing as mproc
 import itertools
 import random
+
+import numpy as np
 import rasterio as rio
+import multiprocessing as mproc
 
-from landiv_blur import helper as lbhelp
-from landiv_blur import io as lbio
-from landiv_blur import io_ as lbio_
-from landiv_blur import processing as lbproc
-from landiv_blur import prepare as lbprep
-from landiv_blur import inference as lbinf
-from landiv_blur import parallel as lbpara
-from landiv_blur.filters import gaussian as lbf_gauss
-from landiv_blur.helper import rasterio_to_numpy_dtype
+from functools import partial
 
-from .conftest import ALL_MAPS, get_file, set_mpc_strategy
+from riogrande import helper as rghelp
+from riogrande import io as rgio
+from riogrande import io_ as rgio_
+from riogrande import prepare as rgprep
+from riogrande import parallel as rgpara
 
-from matplotlib import pyplot as plt
+from convster import processing as csproc
+from convster import prepare as csprep
+from convster import parallel as cspara
+from convster.filters import gaussian as csf_gauss
+
+from .conftest import ALL_MAPS, get_file
+
 
 @ALL_MAPS
 def test_blur_recombination(datafiles, set_mpc_strategy):
@@ -34,13 +32,13 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
     truncate = 3  # property of the gaussian filter
     view_size = (500, 400)
     output_dtype = "uint8"  # data type to use for the blurred arrays
-    blur_params = lbprep.get_blur_params(diameter=_diameter, truncate=truncate)
-    min_border = lbf_gauss.compatible_border_size(sigma=blur_params['sigma'],
+    blur_params = csprep.get_blur_params(diameter=_diameter, truncate=truncate)
+    min_border = csf_gauss.compatible_border_size(sigma=blur_params['sigma'],
                                                   truncate=truncate)
     border = (100, 100)
     # load the data
     ch_map_tif = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
-    ch_map = lbio.load_map(ch_map_tif, indexes=1)
+    ch_map = rgio.load_map(ch_map_tif, indexes=1)
     ch_data = ch_map['data']
     profile = ch_map['orig_profile']
     width = profile['width']
@@ -49,12 +47,12 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
     profile['count'] = 1
     profile['dtype'] = rio.uint8
     # get the categories 
-    categories = lbproc.get_categories(ch_data)
+    categories = csproc.get_categories(ch_data)
     # we partially evaluate the guassian filter to make sure it gets
     # identical parameter everywhere
     # we do not need to pass the diameter to the filter function
     _ = blur_params.pop('diameter')
-    img_filter = partial(lbf_gauss.gaussian, **blur_params)
+    img_filter = partial(csf_gauss.gaussian, **blur_params)
     # we perform the test for each category 
     # print("CATEGORIES", categories)
     for category in categories:
@@ -63,14 +61,14 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
         #This was changed to allow for (1) seleciton of non-sequential lists, (2) lc maps with values [0, 255]
         index = 1
         # perform the blur in a single run
-        blurred_data = lbproc.get_category_data(ch_data,
+        blurred_data = csproc.get_category_data(ch_data,
                                                 category=category,
                                                 img_filter=img_filter,
                                                 as_dtype=output_dtype
                                                 )
         # use multiprocessing and blur block by block
         # first set the parameters for the recombintion task
-        blur_output_file = lbhelp.output_filename(
+        blur_output_file = rghelp.output_filename(
             base_name=blur_partial,
             out_type=f"blur_lct_{category}",
             blur_params=blur_params
@@ -81,7 +79,7 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
             as_dtype=output_dtype
         )
         # now the parameter for the per block blur tasks
-        views, inner_views = lbprep.create_views(view_size=view_size,
+        views, inner_views = rgprep.create_views(view_size=view_size,
                                                  border=border,
                                                  size=(width, height))
         block_params = []
@@ -98,20 +96,20 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
         manager = mproc.Manager()
         blur_q = manager.Queue()
         # get number of workers
-        nbr_workers = lbhelp.get_nbr_workers()
+        nbr_workers = rghelp.get_nbr_workers()
         pool = set_mpc_strategy.Pool(nbr_workers)
         # start the blurred category writer task
         blur_combiner = pool.apply_async(
-            lbpara.combine_blurred_categories,
+            cspara.combine_blurred_categories,
             (blur_output_params, blur_q,)
         )
         # start the block processing
         all_jobs = []
         for bparams in block_params:
             all_jobs.append(pool.apply_async(
-                lbpara.runner_call,
+                rgpara.runner_call,
                 (blur_q,
-                 lbproc.view_blurred,
+                 csproc.view_blurred,
                  bparams)
             ))
         # now lets wait for all of these jobs to finish
@@ -131,13 +129,13 @@ def test_blur_recombination(datafiles, set_mpc_strategy):
 
         # check if tags were set correctly
         with rio.open(blur_output_file) as src:
-            tags = lbio.get_tags(src, bidx=index)
-            bidx = lbio.get_bidx(src, category=category)
+            tags = rgio.get_tags(src, bidx=index)
+            bidx = rgio.get_bidx(src, category=category)
             np.testing.assert_equal(tags['category'], category)
             np.testing.assert_equal(bidx, index)
 
         # now we can read out the tif with the blurred category and compare
-        blurred_cat_map = lbio.load_map(blur_output_file, indexes=index)
+        blurred_cat_map = rgio.load_map(blur_output_file, indexes=index)
         blurred_cat_data = blurred_cat_map['data']
 
         # plt.imshow(blurred_data)
@@ -172,14 +170,14 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
         blur_output_dtype = test_dtype  # blurred maps will be saved in this format
         output_dtype = test_dtype
         normed = True
-        blur_params = lbprep.get_blur_params(diameter=_diameter, truncate=truncate)
-        min_border = lbf_gauss.compatible_border_size(sigma=blur_params['sigma'],
+        blur_params = csprep.get_blur_params(diameter=_diameter, truncate=truncate)
+        min_border = csf_gauss.compatible_border_size(sigma=blur_params['sigma'],
                                                       truncate=truncate)
         border = (50, 50)
         print(f"{min_border=}, {border=}")
         # load the data
         ch_map_tif = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
-        ch_map = lbio.load_map(ch_map_tif, indexes=1)
+        ch_map = rgio.load_map(ch_map_tif, indexes=1)
         ch_data = ch_map['data']
         profile = ch_map['orig_profile']
         width = profile['width']
@@ -188,16 +186,16 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
         profile['count'] = 1
         profile['dtype'] = np.dtype(output_dtype)
         # get the categories
-        categories = lbproc.get_categories(ch_data)
+        categories = csproc.get_categories(ch_data)
         max_entropy_categories = len(categories)
         # we partially evaluate the guassian filter to make sure it gets
         # identical parameter everywhere
         # we do not need to pass the diameter to the filter function
         _ = blur_params.pop('diameter')
-        img_filter = partial(lbf_gauss.gaussian, **blur_params)
+        img_filter = partial(csf_gauss.gaussian, **blur_params)
         filter_params = blur_params.copy()
         filter_params.update(dict(preserve_range=True))
-        entropy_data = lbproc.get_entropy(
+        entropy_data = csproc.get_entropy(
             data=ch_data,
             categories=categories,
             max_entropy_categories=max_entropy_categories,
@@ -210,7 +208,7 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
         )
         # use multiprocessing and blur block by block
         # first set the parameters for the recombintion task
-        entropy_output_file = lbhelp.output_filename(
+        entropy_output_file = rghelp.output_filename(
             base_name=blur_partial,
             out_type=f"entropy_lct",
             blur_params=blur_params
@@ -222,7 +220,7 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
             count=len(categories)
         )
         # now the parameter for the per block blur tasks
-        views, inner_views = lbprep.create_views(view_size=view_size,
+        views, inner_views = rgprep.create_views(view_size=view_size,
                                                  border=border,
                                                  size=(width, height))
         block_params = []
@@ -244,21 +242,21 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
         manager = mproc.Manager()
         entropy_q = manager.Queue()
         # get number of workers
-        nbr_workers = lbhelp.get_nbr_workers()
+        nbr_workers = rghelp.get_nbr_workers()
         # print(f"using {nbr_workers=}")
         pool = set_mpc_strategy.Pool(nbr_workers)
         # start the blurred category writer task
         blur_combiner = pool.apply_async(
-            lbpara.combine_entropy_blocks,
+            cspara.combine_entropy_blocks,
             (entropy_output_params, entropy_q,)
         )
         # start the block processing
         all_jobs = []
         for bparams in block_params:
             all_jobs.append(pool.apply_async(
-                lbpara.runner_call,
+                rgpara.runner_call,
                 (entropy_q,
-                 lbproc.get_entropy_view,
+                 csproc.get_entropy_view,
                  bparams)
             ))
         # now lets wait for all of these jobs to finish
@@ -278,13 +276,13 @@ def test_entropy_recombination(datafiles, set_mpc_strategy):
 
         # check if tags were set correctly
         with rio.open(entropy_output_file) as src:
-            tags = lbio.get_tags(src, bidx=1)
-            bidx = lbio.get_bidx(src, category="entropy")
+            tags = rgio.get_tags(src, bidx=1)
+            bidx = rgio.get_bidx(src, category="entropy")
             np.testing.assert_equal(tags['category'], "entropy")
             np.testing.assert_equal(bidx, 1)
 
         # now we can read out the tif with the blurred category and compare
-        entropy_recomb_map = lbio.load_map(entropy_output_file, indexes=1)
+        entropy_recomb_map = rgio.load_map(entropy_output_file, indexes=1)
         entropy_recomb_data = entropy_recomb_map['data']
 
         # plt.imshow(entropy_data)
@@ -320,13 +318,13 @@ def test_entropy_2_step(datafiles):
         output_dtype = test_dtype  # data type to use for the entropy array
         blur_output_dtype = test_dtype
         normed = True
-        blur_params = lbprep.get_blur_params(diameter=_diameter, truncate=truncate)
-        min_border = lbf_gauss.compatible_border_size(sigma=blur_params['sigma'],
+        blur_params = csprep.get_blur_params(diameter=_diameter, truncate=truncate)
+        min_border = csf_gauss.compatible_border_size(sigma=blur_params['sigma'],
                                                       truncate=truncate)
         border = (50, 50)
         # load the data
         ch_map_tif = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
-        ch_map = lbio.load_map(ch_map_tif, indexes=1)
+        ch_map = rgio.load_map(ch_map_tif, indexes=1)
         ch_data = ch_map['data']
         profile = ch_map['orig_profile']
         width = profile['width']
@@ -335,9 +333,9 @@ def test_entropy_2_step(datafiles):
         profile['count'] = 1
         profile['dtype'] = test_dtype
         # get the categories
-        categories = lbproc.get_categories(ch_data)
+        categories = csproc.get_categories(ch_data)
         # Filter params for both
-        img_filter = lbf_gauss.gaussian
+        img_filter = csf_gauss.gaussian
         filter_params = blur_params.copy()
         filter_params.update(dict(preserve_range=True))
         _ = filter_params.pop('diameter')
@@ -345,7 +343,7 @@ def test_entropy_2_step(datafiles):
         # ###
         # use single step approach
         # ###
-        entropy_output_file = lbhelp.output_filename(
+        entropy_output_file = rghelp.output_filename(
             base_name=blur_partial,
             out_type=f"entropy_lct",
             blur_params=blur_params.copy()
@@ -357,7 +355,7 @@ def test_entropy_2_step(datafiles):
             count=len(categories)
         )
         # now the parameter for the per block blur tasks
-        views, inner_views = lbprep.create_views(view_size=view_size,
+        views, inner_views = rgprep.create_views(view_size=view_size,
                                                  border=border,
                                                  size=(width, height))
         block_params = []
@@ -376,20 +374,20 @@ def test_entropy_2_step(datafiles):
             block_params.append(bparams)
         manager = mproc.Manager()
         entropy_q = manager.Queue()
-        nbr_workers = lbhelp.get_nbr_workers()
+        nbr_workers = rghelp.get_nbr_workers()
         pool = mproc.Pool(nbr_workers)
         # start the blurred category writer task
         blur_combiner = pool.apply_async(
-            lbpara.combine_entropy_blocks,
+            cspara.combine_entropy_blocks,
             (entropy_output_params, entropy_q,)
         )
         # start the block processing
         all_jobs = []
         for bparams in block_params:
             all_jobs.append(pool.apply_async(
-                lbpara.runner_call,
+                rgpara.runner_call,
                 (entropy_q,
-                 lbproc.get_entropy_view,
+                 csproc.get_entropy_view,
                  bparams)
             ))
         # now lets wait for all of these jobs to finish
@@ -410,8 +408,8 @@ def test_entropy_2_step(datafiles):
         ###
         # Now calculate first the map with blurred layers and then the entropy
         ###
-        source = lbio_.Source(path=ch_map_tif)
-        blurred_tif = lbpara.extract_categories(
+        source = rgio_.Source(path=ch_map_tif)
+        blurred_tif = cspara.extract_categories(
             source=source,
             categories=categories,
             output_file=blur_out,
@@ -426,19 +424,19 @@ def test_entropy_2_step(datafiles):
         # get the somewhat weird output filename
         # check if tags were set correctly for the blurred layers
         with rio.open(blurred_tif) as src:
-            tags = lbio.get_tags(src, bidx=1)
-            bidx = lbio.get_bidx(src, category=categories[0])
+            tags = rgio.get_tags(src, bidx=1)
+            bidx = rgio.get_bidx(src, category=categories[0])
             np.testing.assert_equal(tags['category'], categories[0])
             np.testing.assert_equal(bidx, 1)
 
-        blurred_source = lbio_.Source(path=blurred_tif)
+        blurred_source = rgio_.Source(path=blurred_tif)
         for bidx in blurred_source.band_indexes:
-            b = lbio_.Band(source=blurred_source, bidx=bidx)
+            b = rgio_.Band(source=blurred_source, bidx=bidx)
             with rio.open(blurred_source.path, 'r') as src:
                 data = src.read(indexes=bidx)
                 np.testing.assert_equal(data, b.get_data())
 
-        entropy_tif = lbpara.compute_entropy(
+        entropy_tif = cspara.compute_entropy(
             source=blurred_source,
             output_file=entropy_out,
             block_size=view_size,
@@ -449,11 +447,11 @@ def test_entropy_2_step(datafiles):
         )
 
         # now we can read out the tif with the blurred category and compare
-        entropy_map = lbio.load_map(entropy_output_file, indexes=1)
+        entropy_map = rgio.load_map(entropy_output_file, indexes=1)
         entropy_data = entropy_map['data']
 
         # for the 2-step approach
-        entropy_source = lbio_.Source(path=entropy_tif)
+        entropy_source = rgio_.Source(path=entropy_tif)
         # get the entropy band as a object
         eband = entropy_source.get_band(category='entropy')
         entropy_data_2step = eband.get_data()
@@ -487,7 +485,7 @@ def test_interaction_parallel_computation(datafiles, create_blurred_tif):
     dtype_tests = {"uint8": (0, 255),
                    "float32": None}
     for test_dtype, _ in dtype_tests.items():
-        out_source = lbio_.Source(path=create_blurred_tif)
+        out_source = rgio_.Source(path=create_blurred_tif)
         categories = [b.tags['category'] for b in out_source.get_bands()]
         # Pairs
         all_possible_pairs = [list(x) for x in itertools.combinations(categories, r=2)]
@@ -495,7 +493,7 @@ def test_interaction_parallel_computation(datafiles, create_blurred_tif):
         test_pair = [1, 3]
 
         # Interaction (parallel)
-        para_interaction_tif = lbpara.compute_interaction(source=out_source,
+        para_interaction_tif = cspara.compute_interaction(source=out_source,
                                                           output_file=str(datafiles / f'{test_dtype}_interact_out.tif'),
                                                           block_size=(500, 500),
                                                           categories=test_pair,
@@ -504,15 +502,215 @@ def test_interaction_parallel_computation(datafiles, create_blurred_tif):
                                                           output_dtype=test_dtype,                                                          standardize=True,
                                                           normed=True,
                                                           verbose=False)
-        int_band = lbio_.Band(source=lbio_.Source(path=para_interaction_tif), bidx=1)
+        int_band = rgio_.Band(source=rgio_.Source(path=para_interaction_tif), bidx=1)
         para_interaction_data = int_band.get_data()
 
         # Interaction (single process
         band_list = [out_source.get_band(category=c) for c in test_pair]
         data_array = [b.get_data() for b in band_list]
-        interaction_data = lbproc.compute_interaction(data_arrays=data_array,
+        interaction_data = csproc.compute_interaction(data_arrays=data_array,
                                                       input_dtype=np.uint8,
                                                       normed=True,
                                                       standardize=True,
                                                       output_dtype=test_dtype)
         np.testing.assert_array_equal(para_interaction_data, interaction_data)
+
+
+@ALL_MAPS
+def test_apply_filter(datafiles):
+    """Test the parallel application of a filter to one or serveral bands
+    """
+    blur_para = str(datafiles / 'blur_para.tif')
+    blur_single = str(datafiles / 'blur_single.tif')
+    bands_out = str(datafiles / 'bands_out.tif')
+    diameter = 1000  # this is in meter
+    scale = 100  # meter per pixel
+    _diameter = diameter / scale
+    truncate = 3  # property of the gaussian filter
+    block_size = (500, 400)
+    output_dtype = np.uint8  # data type to use for the blurred arrays
+    blur_params = csprep.get_blur_params(diameter=_diameter, truncate=truncate)
+    img_filter=csf_gauss.gaussian
+    filter_params = blur_params.copy()
+    _ = filter_params.pop('diameter')
+    ch_map_tif = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
+    categories = [1,2,3]
+    # compute in one go
+    blurred_tif = cspara.extract_categories(
+        source=str(ch_map_tif),
+        categories=categories,
+        output_file=blur_single,
+        img_filter=img_filter,
+        filter_params=filter_params,
+        filter_output_range=(0,1),
+        output_params=dict(
+            as_dtype=output_dtype,
+            output_range=output_dtype
+        ),
+        block_size=block_size,
+        compress=True
+    )
+    # now in two steps:
+    # first extract categories
+    bands_tif = cspara.extract_categories(
+        source=str(ch_map_tif),
+        categories=categories,
+        output_file=bands_out,
+        img_filter=None,
+        filter_params=filter_params,
+        output_dtype=output_dtype,
+        block_size=block_size,
+        verbose=True,
+        compress=True
+    )
+    # apply the filter in parallel
+    blurred_para = cspara.apply_filter(source=bands_tif,
+                                       output_file=blur_para,
+                                       block_size=block_size,
+                                       bands=None,
+                                       data_as_dtype=np.uint8,
+                                       img_filter=img_filter,
+                                       filter_params=filter_params,
+                                       filter_output_range=(0.,1.),
+                                       output_dtype=output_dtype,
+                                       verbose=True)
+    for cat in categories:
+        b_nope = rgio_.Band(source=rgio_.Source(path=bands_tif),
+                            bidx=cat)
+        b_twostep = rgio_.Band(source=rgio_.Source(path=blurred_para),
+                               bidx=cat)
+        b_single = rgio_.Band(source=rgio_.Source(path=blurred_tif),
+                              bidx=cat)
+        b_nope.import_tags()
+        b_twostep.import_tags()
+        b_single.import_tags()
+        np.testing.assert_equal(b_twostep.get_data(), b_single.get_data())
+
+@ALL_MAPS
+def test_extract_categories(datafiles):
+    """Make sure the extract categories works as expected
+    """
+    verbose = True
+    landcover_map = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
+    lct_source = rgio_.Source(path=landcover_map)
+    source_profile = lct_source.import_profile()
+    source_band = lct_source.get_band(bidx=1)
+    print(f"{source_profile=}")
+    # extract categories without applying a filter
+    to_dtype="uint8"
+    categories = [1,2,3,4,5]
+    category_tif = cspara.extract_categories(
+        source=lct_source,
+        categories=categories,
+        output_file=str(datafiles / 'category_out.tif'),
+        output_dtype=to_dtype,
+        block_size=(500, 500),
+        compress = True,
+        output_params = dict(
+            nodata=0,
+            dtype=to_dtype
+        ),
+    )
+    category_source = rgio_.Source(category_tif)
+    assert len(category_source.get_bands()) == len(categories)
+    source_data = source_band.get_data()
+    for cat in categories:
+        # check for each category whether the extraction
+        # is identical to the real data
+        cat_band = category_source.get_band(category=cat)
+        cat_data = cat_band.get_data()
+        _cat_data = np.where(cat_data == 255, 1, 0)
+        _category_data = np.where(source_data==cat, 1, 0)
+        np.testing.assert_equal(_cat_data, _category_data)
+
+    # check nodata handling
+    # creat an output file (with changed nodata)
+    to_dtypes = ["float32", "int16", "uint8"]
+    nodatas = [np.nan, 0, None]
+    for nodata, to_dtype in zip(nodatas, to_dtypes):
+        tmp_map = str(datafiles / 'bands_out.tif')
+        tmp_source = rgio_.Source(path=tmp_map)
+        tmp_profile = source_profile.copy()
+        tmp_profile['nodata'] = nodata
+        tmp_profile['dtype'] = to_dtype
+        tmp_source.profile = tmp_profile
+        tmp_source.init_source(overwrite=True)
+        # sanity check for Source.init_source resp. Source.open
+        with rio.open(tmp_source.path, 'r') as src:
+            _profile = src.profile.copy()
+        np.testing.assert_equal(_profile['nodata'], nodata)
+
+        tmp_band = rgio_.Band(source=tmp_source, bidx=1)
+        # write out data as 
+        tmp_band.set_data(source_band.get_data().astype(to_dtype))
+        filter_params = dict(
+            sigma = 100,
+            truncate = 3
+        )
+        blurred_tif = cspara.extract_categories(
+            source=tmp_source,
+            categories=[1,2,3,4,5],
+            output_file=str(datafiles / 'blur_out.tif'),
+            img_filter=csf_gauss.gaussian,
+            filter_params=filter_params,
+            output_dtype=to_dtype,
+            block_size=(500, 500),
+            compress = True,
+            output_params = dict(
+                nodata=nodata,
+                dtype=to_dtype
+            ),
+        )
+        out_source = rgio_.Source(path=blurred_tif)
+        out_profile = out_source.import_profile()
+        print(f"{out_profile=}")
+        np.testing.assert_equal(out_profile['nodata'], nodata)
+        # We need to map GDAL to numpy datatypes
+        assert out_profile['dtype'] == to_dtype
+        #np.testing.assert_equal(rasterio_to_numpy_dtype(out_profile['dtype']), to_dtype)
+
+@ALL_MAPS
+def test_reduced_mask(datafiles):
+    """Compute a mask from multiple bands in one go and then in parallel
+    """
+    ch_map_tif = get_file(pattern="Switzerland_CLC_*.tif", datafiles=datafiles)
+    source = rgio_.Source(path=ch_map_tif)
+    blur_out = str(datafiles / 'blur_out.tif')
+    # create the blurred bands
+    img_filter = csf_gauss.gaussian
+    output_dtype = np.uint8
+    diameter = 5000  # this is in meter
+    scale = 100  # meter per pixel
+    truncate = 3
+    view_size = (500, 400)
+    categories = [1, 2, 3, 4, 5]
+    _diameter = diameter / scale
+    blur_params = csprep.get_blur_params(diameter=_diameter, truncate=truncate)
+    filter_params = blur_params.copy()
+    _ = filter_params.pop('diameter')
+    blurred_tif = cspara.extract_categories(
+        source=source,
+        categories=categories,
+        output_file=blur_out,
+        img_filter=img_filter,
+        filter_params=filter_params,
+        output_dtype=output_dtype,
+        block_size=view_size,
+        compress=True
+    )
+    blurr_source = rgio_.Source(path=blurred_tif)
+    initial_mask = blurr_source.get_mask()
+    # get the mask loading the entire dataset
+    with blurr_source.data_reader(mode='r') as read:
+        dataset = read()
+    # print(f"{dataset.shape=}")
+    mask = rghelp.reduced_mask(array=dataset)
+    # print(f"{mask=}")
+    rgpara.compute_mask(source=blurr_source, block_size=view_size)
+    updated_mask = blurr_source.get_mask()
+    # as get_mask returns [0, 255] mask and mask produces [0, 1] we need to account for that
+    # it is important that > 0 is Valid data and needs to be equal
+    updated_mask = np.divide(updated_mask, 255)
+    # print(f"UNIQUE VALUES: \n mask: {np.unique(mask)}\n updated_mask: {np.unique(updated_mask)}")
+    np.testing.assert_array_equal(mask, updated_mask)
+    assert not np.array_equal(initial_mask, updated_mask)
