+++ tests/test_linfit/test_inference.py
@@ -1,18 +1,19 @@
-import pytest
-
 import numpy as np
 import rasterio as rio
 from numpy.random import Generator, PCG64
 
 from pydataset import data as pydata
 
-from landiv_blur import helper as lbhelp
-from riogrande import io as lbio
-from riogrande import io_ as lbio_
-from landiv_blur import parallel as lbpara
-from linfit import inference as lbinf
+from riogrande import helper as rghelp
 
-from .conftest import ALL_MAPS, get_file, set_mpc_strategy
+from riogrande import io as rgio
+from riogrande import io_ as rgio_
+from riogrande import parallel as rgpara
+
+from linfit import inference as lfinf
+from linfit import parallel as lfpara
+
+from .conftest import ALL_MAPS, get_file
 
 
 @ALL_MAPS
@@ -26,17 +27,17 @@ def test_preparation(datafiles, create_blurred_tif, set_mpc_strategy):
     print(f"{ndvi_map=}")
 
     landcover_map = str(datafiles / 'lct_coreged.tif')
-    lbio._coregister_raster(_landcover_map, ndvi_map, output=landcover_map)
+    rgio._coregister_raster(_landcover_map, ndvi_map, output=landcover_map)
     # work with strings first
-    lbinf.prepare_predictors(ndvi_map,
+    lfinf.prepare_predictors(ndvi_map,
                              landcover_map,
                              include_intercept=True,)
     # now same with Band objects
-    lct_source = lbio_.Source(path=landcover_map)
-    lct_band = lbio_.Band(source=lct_source, bidx=1)
-    ndvi_source = lbio_.Source(path=ndvi_map)
-    ndvi_band = lbio_.Band(source=ndvi_source, bidx=1)
-    lbinf.prepare_predictors(ndvi_band,
+    lct_source = rgio_.Source(path=landcover_map)
+    lct_band = rgio_.Band(source=lct_source, bidx=1)
+    ndvi_source = rgio_.Source(path=ndvi_map)
+    ndvi_band = rgio_.Band(source=ndvi_source, bidx=1)
+    lfinf.prepare_predictors(ndvi_band,
                              lct_band,
                              include_intercept=True,)
 
@@ -48,14 +49,14 @@ def test_optimal_weights_test_data():
     # column 4 should be 1 and the others close to 0
     X = iris[iris.columns[:4]].values
     y = iris[iris.columns[3]].values
-    b = np.round(lbinf.get_optimal_weights(X, y), 6)
+    b = np.round(lfinf.get_optimal_weights(X, y), 6)
     np.testing.assert_array_equal(b, np.array([0, 0, 0, 1]))
     # now we add some white noise to column 4 (as predictor):
     X_noisy = np.copy(X)
     rng = Generator(PCG64(seed=33))
     X_noisy[:, 3] += rng.standard_normal(X_noisy.shape[0]) * 0.02
     # this should affect the quality of the fit (so we round to first digit:
-    b_noisy = np.round(lbinf.get_optimal_weights(X_noisy, y), 1)
+    b_noisy = np.round(lfinf.get_optimal_weights(X_noisy, y), 1)
     np.testing.assert_array_equal(b_noisy, np.array([0, 0, 0, 1]))
 
 
@@ -67,9 +68,9 @@ def test_weights_computations_test_data():
     # column 4 should be 1 and the others close to 0
     X = iris[iris.columns[:4]].values
     y = iris[iris.columns[3]].values
-    b = np.round(lbinf.get_optimal_weights(X, y), 6)
+    b = np.round(lfinf.get_optimal_weights(X, y), 6)
     # now use the linear regression estimation
-    reg = lbinf.get_approx_weights(X, y)
+    reg = lfinf.get_approx_weights(X, y)
     b_approx = np.round(reg.coef_)
     # print(f"\n{b=}\n{b_approx=}")
     # print(reg.intercept_)
@@ -79,8 +80,8 @@ def test_weights_computations_test_data():
     X_noisy = np.copy(X)
     X_noisy[:, 3] += rng.standard_normal(X_noisy.shape[0]) * 0.02
     # this should affect the quality of the fit (so we round to first digit:
-    b_noisy = np.round(lbinf.get_optimal_weights(X_noisy, y), 1)
-    reg_noisy = lbinf.get_approx_weights(X_noisy, y)
+    b_noisy = np.round(lfinf.get_optimal_weights(X_noisy, y), 1)
+    reg_noisy = lfinf.get_approx_weights(X_noisy, y)
     b_noisy_approx = np.round(reg_noisy.coef_)
     # print(f"{b_noisy=}\n{b_noisy_approx=}")
     np.testing.assert_array_equal(b_noisy, b_noisy_approx)
@@ -97,7 +98,7 @@ def test_optimal_weights_example_data(datafiles, create_blurred_tif):
     # scale it down to 100x100m (from 30x30)
 
     ndvi_map = str(datafiles / 'lct_coreged.tif')
-    lbio._coregister_raster(_ndvi_map, landcover_map, output=ndvi_map)
+    rgio._coregister_raster(_ndvi_map, landcover_map, output=ndvi_map)
 
     # create a mask for ndvi_map masking the nan's
     with rio.open(ndvi_map, 'r+') as src:
@@ -105,28 +106,28 @@ def test_optimal_weights_example_data(datafiles, create_blurred_tif):
         mask = np.where(np.isnan(data), 0, 255)
         src.write_mask(mask)
 
-    blurred_source = lbio_.Source(path=create_blurred_tif)
+    blurred_source = rgio_.Source(path=create_blurred_tif)
     predictors = blurred_source.get_bands()
     # choose the write mask
     for pred in predictors:
         pred.set_mask_reader(use='source')
 
-    X, y = lbinf.prepare_predictors(ndvi_map,
+    X, y = lfinf.prepare_predictors(ndvi_map,
                                     *predictors,
                                     include_intercept=False,
                                     verbose=True,
                                     )
-    b = np.round(lbinf.get_optimal_weights(X, y), 6)
-    reg = lbinf.get_approx_weights(X, y, fit_intercept=False)
+    b = np.round(lfinf.get_optimal_weights(X, y), 6)
+    reg = lfinf.get_approx_weights(X, y, fit_intercept=False)
     b_approx = np.round(reg.coef_, 6)
     # print(f"\n{b=}\n{b_approx=}\n")
     np.testing.assert_allclose(b, b_approx, rtol=1e-05)
     # now check when ingnoring the intercept in both cases
-    X, y = lbinf.prepare_predictors(ndvi_map,
+    X, y = lfinf.prepare_predictors(ndvi_map,
                                     *predictors,
                                     include_intercept=False,)
-    b = np.round(lbinf.get_optimal_weights(X, y), 6)
-    reg = lbinf.get_approx_weights(X, y, fit_intercept=False)
+    b = np.round(lfinf.get_optimal_weights(X, y), 6)
+    reg = lfinf.get_approx_weights(X, y, fit_intercept=False)
     b_approx = np.round(reg.coef_, 6)
     # print(f"\n{b=}\n{b_approx=}\n")
     np.testing.assert_allclose(b, b_approx, rtol=1e-05)
@@ -143,18 +144,18 @@ def test_transposed_prod_example_data(datafiles, create_blurred_tif,
 
     # scale it down to 100x100m (from 30x30)
     ndvi_map = str(datafiles / 'lct_coreged.tif')
-    lbio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
+    rgio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
 
-    lct_source = lbio_.Source(path=landcover_map)
-    ndvi_source = lbio_.Source(path=ndvi_map)
-    blurred_source = lbio_.Source(path=create_blurred_tif)
+    lct_source = rgio_.Source(path=landcover_map)
+    ndvi_source = rgio_.Source(path=ndvi_map)
+    blurred_source = rgio_.Source(path=create_blurred_tif)
     # set the mask
-    lbpara.compute_mask(source=blurred_source,
+    rgpara.compute_mask(source=blurred_source,
                         block_size=(500, 500),
                         nodata=0,
                         logic='all')
     # create the inputs
-    response = lbio_.Band(source=lbio_.Source(path=ndvi_map))
+    response = rgio_.Band(source=rgio_.Source(path=ndvi_map))
     predictors = blurred_source.get_bands()
     # Each band should use the dataset mask:
     for pred_band in predictors:
@@ -167,16 +168,16 @@ def test_transposed_prod_example_data(datafiles, create_blurred_tif,
         mask = np.where(np.isnan(data), 0, 255)
         src.write_mask(mask)
 
-    X, y = lbinf.prepare_predictors(response,
+    X, y = lfinf.prepare_predictors(response,
                                     *predictors,
                                     include_intercept=False,
                                     verbose=True,
                                     )
     tpX = X.T @ X
     # calculate it per predictor columns
-    selector = lbinf.prepare_selector(response,
+    selector = lfinf.prepare_selector(response,
                                       *predictors)
-    transprodX = lbinf.transposed_product(predictors,
+    transprodX = lfinf.transposed_product(predictors,
                                           view=None,
                                           selector=selector,
                                           as_dtype="float64")
@@ -193,15 +194,15 @@ def test_extra_masking_band(datafiles, create_blurred_tif, set_mpc_strategy):
 
     # scale it down to 100x100m (from 30x30)
     ndvi_map = str(datafiles / 'lct_coreged.tif')
-    lbio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
-    blurred_source = lbio_.Source(path=create_blurred_tif)
+    rgio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
+    blurred_source = rgio_.Source(path=create_blurred_tif)
     # set the mask
-    lbpara.compute_mask(source=blurred_source,
+    rgpara.compute_mask(source=blurred_source,
                         block_size=(500, 500),
                         nodata=0,
                         logic='all')
     # create the inputs
-    response = lbio_.Band(source=lbio_.Source(path=ndvi_map))
+    response = rgio_.Band(source=rgio_.Source(path=ndvi_map))
     predictors = blurred_source.get_bands()
     # Each band should use the dataset mask:
     for pred_band in predictors:
@@ -210,17 +211,17 @@ def test_extra_masking_band(datafiles, create_blurred_tif, set_mpc_strategy):
     # Create extra masking band
     resp_profile = response.source.import_profile()
     tmp_map = str(datafiles / 'extra_mask_band.tif')
-    tmp_source = lbio_.Source(path=tmp_map)
+    tmp_source = rgio_.Source(path=tmp_map)
     tmp_profile = resp_profile.copy()
     tmp_profile['nodata'] = 0
     tmp_profile['dtype'] = np.uint8
     tmp_source.profile = tmp_profile
     tmp_source.init_source(overwrite=True)
-    extra_masking_band = lbio_.Band(source=tmp_source, bidx=1)
+    extra_masking_band = rgio_.Band(source=tmp_source, bidx=1)
     # write out data as (mask all)
     extra_mask_data = np.full(shape=response.shape, fill_value=0, dtype=np.uint8)
     extra_masking_band.set_data(data=extra_mask_data)
-    selector = lbinf.prepare_selector(
+    selector = lfinf.prepare_selector(
         response,
         *predictors,
         extra_masking_band=extra_masking_band)
@@ -228,10 +229,10 @@ def test_extra_masking_band(datafiles, create_blurred_tif, set_mpc_strategy):
     # extra mask none and check that it has no influence
     extra_mask_data = np.full(shape=response.shape, fill_value=255, dtype=np.uint8)
     extra_masking_band.set_data(data=extra_mask_data)
-    selector_wo = lbinf.prepare_selector(
+    selector_wo = lfinf.prepare_selector(
         response,
         *predictors,)
-    selector = lbinf.prepare_selector(
+    selector = lfinf.prepare_selector(
         response,
         *predictors,
         extra_masking_band=extra_masking_band)
@@ -245,7 +246,7 @@ def test_transposed_prod_blurred_example_data(datafiles, create_blurred_tif):
     # get the response data
     _ndvi_map = get_file(pattern="Switzerland_NDVI_*.tif", datafiles=datafiles)
     # get (compute) the blurred bands
-    blurred_source = lbio_.Source(path=create_blurred_tif)
+    blurred_source = rgio_.Source(path=create_blurred_tif)
     # parameter setting
     view = None  # use the full maps
     include_intercept = True
@@ -254,7 +255,7 @@ def test_transposed_prod_blurred_example_data(datafiles, create_blurred_tif):
     sigma = 10
     # scale it down to 100x100m (from 30x30)
     ndvi_map = str(datafiles / 'lct_coreged.tif')
-    lbio._coregister_raster(_ndvi_map, blurred_source.path, output=ndvi_map)
+    rgio._coregister_raster(_ndvi_map, blurred_source.path, output=ndvi_map)
     # create a mask for ndvi_map masking the nan's
     with rio.open(ndvi_map, 'r+') as src:
         data = src.read(indexes=1)
@@ -268,18 +269,18 @@ def test_transposed_prod_blurred_example_data(datafiles, create_blurred_tif):
     for pred in predictors:
         pred.set_mask_reader(use='source')
 
-    selector = lbinf.prepare_selector(response,
+    selector = lfinf.prepare_selector(response,
                                       *predictors,
                                       verbose=verbose)
-    riow = lbhelp.view_to_window(view)
-    X = lbinf.init_X(predictors,
+    riow = rghelp.view_to_window(view)
+    X = lfinf.init_X(predictors,
                      selector=selector,
                      window=riow,
                      include_intercept=include_intercept,
                      as_dtype=as_dtype
                      )
 
-    lbinf.populate_X(X=X,
+    lfinf.populate_X(X=X,
                      predictors=predictors,
                      window=riow,
                      selector=selector,
@@ -306,7 +307,7 @@ def test_optimal_beta(datafiles, create_blurred_tif):
     # scale it down to 100x100m (from 30x30)
 
     ndvi_map = str(datafiles / 'ndvi_coreged.tif')
-    lbio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
+    rgio._coregister_raster(_ndvi_map, landcover_map, output=str(ndvi_map))
     # create a mask for ndvi_map masking the nan's
     with rio.open(ndvi_map, 'r+') as src:
         data = src.read(indexes=1)
@@ -315,7 +316,7 @@ def test_optimal_beta(datafiles, create_blurred_tif):
 
     # create the inputs
     response = ndvi_map
-    blurred_source = lbio_.Source(path=create_blurred_tif)
+    blurred_source = rgio_.Source(path=create_blurred_tif)
     predictors = blurred_source.get_bands()
     # choose the write mask
     for pred in predictors:
@@ -324,7 +325,7 @@ def test_optimal_beta(datafiles, create_blurred_tif):
     # test for intercept set and not set
     intercept_switch = [True, False]
     for isetup in intercept_switch:
-        X, y = lbinf.prepare_predictors(response,
+        X, y = lfinf.prepare_predictors(response,
                                         *predictors,
                                         include_intercept=isetup,
                                         verbose=True,
@@ -335,15 +336,15 @@ def test_optimal_beta(datafiles, create_blurred_tif):
         # calculate optimal betas
         betas = Y @ X.T @ y
         # calculate it per predictor columns
-        selector = lbinf.prepare_selector(response,
+        selector = lfinf.prepare_selector(response,
                                           *predictors)
-        tpX_col = lbinf.transposed_product(predictors,
+        tpX_col = lfinf.transposed_product(predictors,
                                            view=None,
                                            include_intercept=isetup,
                                            selector=selector,
                                            as_dtype=as_dtype)
         Y_col = np.linalg.inv(tpX_col)
-        betas_col = lbinf.get_optimal_weights_source(Y=Y,
+        betas_col = lfinf.get_optimal_weights_source(Y=Y,
                                                      response=response,
                                                      predictors=predictors,
                                                      view=None,
